-Maven repository for ITUBIDB(maven.test.itu.edu.tr for now and will be change to maven.itu.edu.tr) have been created and implemented in pom.xml file. - Package name have changed to tr.edu.itu.cavabunga and cavabunga.jar manual dependency removed.
