+++ pom.xml
@@ -3,12 +3,12 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>tr.edu.itu</groupId>
+	<groupId>tr.edu.itu.cavabunga</groupId>
 	<artifactId>cavabunga-server</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
+	<version>1.0.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
-	<name>cavabunga</name>
+	<name>cavabunga-server</name>
 	<description>Caldav / Carddav Server Implementation</description>
 
 	<parent>
@@ -84,13 +84,25 @@
 			<version>3.7</version>
 		</dependency>
 		<dependency>
-			<groupId>tr.edu.itu</groupId>
+			<groupId>tr.edu.itu.cavabunga</groupId>
 			<artifactId>cavabunga-lib</artifactId>
-			<version>1.0.1</version>
-			<scope>system</scope>
-			<systemPath>${project.basedir}/src/main/resources/cavabunga-lib.jar</systemPath>
+			<version>1.0.2-SNAPSHOT</version>
 		</dependency>
 	</dependencies>
+
+	<repositories>
+		<repository>
+			<releases>
+				<enabled>true</enabled>
+				<updatePolicy>always</updatePolicy>
+				<checksumPolicy>fail</checksumPolicy>
+			</releases>
+			<id>cavabunga</id>
+			<name>cavabunga-lib</name>
+			<url>http://maven.test.itu.edu.tr</url>
+			<layout>default</layout>
+		</repository>
+	</repositories>
 
 	<build>
 		<plugins>
