--- src/main/java/edu/itu/cavabunga/CavabungaApplication.java
@@ -1,12 +1,0 @@
-package edu.itu.cavabunga;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CavabungaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CavabungaApplication.class, args);
-	}
-}
