--- src/main/java/edu/itu/cavabunga/repository/ParameterRepository.java
@@ -1,12 +1,0 @@
-package edu.itu.cavabunga.repository;
-
-import edu.itu.cavabunga.lib.entity.Parameter;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface ParameterRepository extends JpaRepository<Parameter, Long> {
-    Optional<Parameter> findById(Long Id);
-}
