+++ src/main/java/tr/edu/itu/cavabunga/server/configuration/SwaggerConfig.java
@@ -1,0 +1,57 @@
+package tr.edu.itu.cavabunga.server.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurationSupport;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class SwaggerConfig extends WebMvcConfigurationSupport {
+    @Bean
+    public Docket productApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                    .apis(RequestHandlerSelectors.any())
+                    .paths(PathSelectors.any())
+                        //.paths(regex("/*"))
+                    .build()
+                .pathMapping("/")
+                .apiInfo(metaData());
+    }
+    private ApiInfo metaData() {
+
+        return new ApiInfo(
+
+                "Cavabunga CORE REST API",
+
+                "Restful api for cavabunga calendar server",
+
+                "1.0",
+
+                "Terms of service",
+
+                "Apache License Version 2.0",
+
+                "https://www.apache.org/licenses/LICENSE-2.0",
+
+                ""
+        );
+    }
+
+
+    @Override
+    protected void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("swagger-ui.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
+
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+    }
+}
