+++ src/main/java/tr/edu/itu/cavabunga/server/service/ParticipantServiceImpl.java
@@ -1,0 +1,90 @@
+package tr.edu.itu.cavabunga.server.service;
+
+import tr.edu.itu.cavabunga.lib.entity.Participant;
+import tr.edu.itu.cavabunga.lib.entity.participant.ParticipantType;
+import tr.edu.itu.cavabunga.lib.factory.ParticipantFactory;
+import tr.edu.itu.cavabunga.server.repository.ParticipantRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
+
+/**
+ * {@inheritDoc}
+ */
+@Service
+public class ParticipantServiceImpl implements ParticipantService {
+
+    private ParticipantFactory participantFactory;
+
+    private ParticipantRepository participantRepository;
+
+    /**
+     * constructor for dependency injection
+     *
+     * @param participantFactory to inject ParticipantFactory
+     * @param participantRepository to inject ParticipantRepository
+     */
+    @Autowired
+    public ParticipantServiceImpl(
+            ParticipantFactory participantFactory,
+            ParticipantRepository participantRepository
+    ) {
+        this.participantFactory = participantFactory;
+        this.participantRepository = participantRepository;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Participant createParticipant(String userName, ParticipantType participantType){
+        Participant newParticipant = participantFactory.createParticipant(participantType);
+        newParticipant.setUserName(userName);
+        //participantRepository.save(newParticipant); //TODO: no need to create just create, while you saving: it will give error "New participant cannot have id field"
+        return newParticipant;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Transactional
+    public void saveParticipant(Participant participant){
+        participantRepository.save(participant);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Optional<Participant> getParticipantByUserName(String user_name){
+        return participantRepository.findByUserName(user_name);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public Optional<Participant> getParticipantById(Long id){
+        return participantRepository.findById(id);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public List<Participant> getAllParticipant(){
+        return participantRepository.findAll();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void deleteParticipantById(Long id){
+        participantRepository.delete(id);
+    }
+}
