+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
 spring.jpa.hibernate.ddl-auto=create
 spring.datasource.url=jdbc:mysql://127.0.0.1/test
-spring.datasource.username=root
+spring.datasource.username=cavabunga
+spring.datasource.password=password
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
