--- src/test/java/edu/itu/cavabunga/CavabungaApplicationTest.java
@@ -1,17 +1,0 @@
-package edu.itu.cavabunga;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.junit.Assert.*;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class CavabungaApplicationTest {
-
-    @Test
-    public void contextLoads() {
-    }
-}
