+++ src/test/java/tr/edu/itu/cavabunga/server/CavabungaApplicationTest.java
@@ -1,0 +1,17 @@
+package tr.edu.itu.cavabunga.server;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.Assert.*;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class CavabungaApplicationTest {
+
+    @Test
+    public void contextLoads() {
+    }
+}
